Support reading 64 bit stream versions
Allow writing 64bit LogRecords so that a 3.9.4 node can replicate off a 4.0.0 node.

Add ChunkVersions for Original, Unaligned and Aligned, and add support
for Aligned ChunkVersions.

The index will now write V3 index entries, supporting 64 bit version numbers.
IndexEntry32Bit and IndexEntry64Bit have been renamed to IndexEntryV1 and V2.

Update Commit- and PrepareLogRecord to support reading 64bit ExpectedVersions.
When reading long.MaxValue expected version on a V1 record, convert it to int.MaxValue.
Do the same conversion when reading $tb that is set to long.MaxValue in v1 metadata records.

Change EventNumber, TruncateBefore and MaxCount from int to long.

Add support for reading V3 PosMaps which include lengthOffset.
Use lengthOffset from PosMap to calculate the next position when reading records.

Add V1 client dispatcher and tcp commands from 4.0.0.

+++ src/EventStore.Core.Tests/Helpers/TestFixtureWithExistingEvents.cs
@@ -377,7 +377,7 @@ namespace EventStore.Core.Tests.Helpers
                 _writesQueue.Enqueue(message);
         }
 
-        private void ProcessWrite<T>(IEnvelope envelope, Guid correlationId, string streamId, int expectedVersion, Event[] events, Func<int, int, T> writeEventsCompleted, T wrongExpectedVersionResponse, long[] positions = null, int? commitPosition = null) where T : Message
+        private void ProcessWrite<T>(IEnvelope envelope, Guid correlationId, string streamId, long expectedVersion, Event[] events, Func<int, int, T> writeEventsCompleted, T wrongExpectedVersionResponse, long[] positions = null, int? commitPosition = null) where T : Message
         {
             if (positions == null)
             {
