+++ src/EventStore.Core.Tests/Services/PersistentSubscription/OutstandingMessageCacheTests.cs
@@ -80,14 +80,14 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription
         public void lowest_on_empty_cache_returns_min()
         {
             var cache = new OutstandingMessageCache();
-            Assert.AreEqual(int.MinValue, cache.GetLowestPosition());
+            Assert.AreEqual(long.MinValue, cache.GetLowestPosition());
         }
         [Test]
         public void get_expired_messages_returns_min_value_on_empty_cache()
         {
             var cache = new OutstandingMessageCache();
             Assert.AreEqual(0, cache.GetMessagesExpiringBefore(DateTime.Now).Count());
-            Assert.AreEqual(int.MinValue, cache.GetLowestPosition());
+            Assert.AreEqual(long.MinValue, cache.GetLowestPosition());
         }
 
         [Test]
