+++ src/EventStore.Projections.Core.Tests/Services/checkpoint_tag/checkpoint_tag_by_event_type_index_positions_when_updating.cs
@@ -10,7 +10,7 @@ namespace EventStore.Projections.Core.Tests.Services.checkpoint_tag
     public class checkpoint_tag_by_event_type_index_positions_when_updating
     {
         private readonly CheckpointTag _a1b1 = CheckpointTag.FromEventTypeIndexPositions(
-            1, new TFPos(100, 50), new Dictionary<string, int> {{"a", 1}, {"b", 1}});
+            1, new TFPos(100, 50), new Dictionary<string, long> {{"a", 1}, {"b", 1}});
 
         [Test]
         public void updated_tf_only_position_is_correct()
